refactor(CHAT): 채팅창 부분 TALK 분류
